Move CommonServiceLocator to seperate project.
- Rename SolrNet nuget to SolrNet.Core
- Update all nuspec to depend on SolrNet.Core
