+++ AutofacContrib.SolrNet/AutofacContrib.SolrNet.nuspec
@@ -12,7 +12,7 @@
     <description>Autofac module for SolrNet</description>
     <language>en-US</language>
     <dependencies>
-      <dependency id="SolrNet" version="[0.7.1]" />
+      <dependency id="SolrNet.Core" version="[0.7.1]" />
       <dependency id="Autofac" version="4.6.1" />
     </dependencies>
   </metadata>
