+++ CommonServiceLocator.SolrNet.Tests/packages.config
@@ -1,0 +1,10 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="CommonServiceLocator" version="1.3" targetFramework="net46" />
+  <package id="xunit" version="2.2.0" targetFramework="net46" />
+  <package id="xunit.abstractions" version="2.0.1" targetFramework="net46" />
+  <package id="xunit.assert" version="2.2.0" targetFramework="net46" />
+  <package id="xunit.core" version="2.2.0" targetFramework="net46" />
+  <package id="xunit.extensibility.core" version="2.2.0" targetFramework="net46" />
+  <package id="xunit.extensibility.execution" version="2.2.0" targetFramework="net46" />
+</packages>
