+++ CommonServiceLocator.SolrNet/CommonServiceLocator.SolrNet.nuspec
@@ -1,25 +1,24 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <package xmlns="http://schemas.microsoft.com/packaging/2010/07/nuspec.xsd">
   <metadata>
-    <id>SolrNet.Windsor</id>
+    <id>SolrNet</id>
     <version>0.7.1</version>
-    <title>SolrNet.Windsor</title>
+    <title>SolrNet</title>
     <authors>Mauricio Scheffer and contributors</authors>
     <owners>Mauricio Scheffer</owners>
     <licenseUrl>https://raw.githubusercontent.com/SolrNet/SolrNet/master/license.txt</licenseUrl>
     <projectUrl>https://github.com/SolrNet/SolrNet</projectUrl>
     <requireLicenseAcceptance>false</requireLicenseAcceptance>
-    <description>Windsor facility for SolrNet</description>
+    <description>Apache Solr client with buildin Dependency Injection.</description>
     <language>en-US</language>
     <dependencies>
+      <dependency id="CommonServiceLocator" version="1.3" />
       <dependency id="SolrNet.Core" version="[0.7.1]" />
-      <dependency id="Castle.Core" version="4.1.1" />
-      <dependency id="Castle.Windsor" version="4.0.0" />
     </dependencies>
   </metadata>
   <files>
-    <file src="bin\release\Castle.Facilities.SolrNetIntegration.dll" target="lib\net46\Castle.Facilities.SolrNetIntegration.dll" />
-    <file src="bin\release\Castle.Facilities.SolrNetIntegration.pdb" target="lib\net46\Castle.Facilities.SolrNetIntegration.pdb" />
-    <file src="bin\release\Castle.Facilities.SolrNetIntegration.XML" target="lib\net46\Castle.Facilities.SolrNetIntegration.xml" />
+    <file src="bin\release\CommonServiceLocator.SolrNet.dll" target="lib\net46\CommonServiceLocator.SolrNet.dll" />
+    <file src="bin\release\CommonServiceLocator.SolrNet.pdb" target="lib\net46\CommonServiceLocator.SolrNet.pdb" />
+    <file src="bin\release\CommonServiceLocator.SolrNet.XML" target="lib\net46\CommonServiceLocator.SolrNet.xml" />
   </files>
 </package>
