+++ CommonServiceLocator.SolrNet/Container.cs
