+++ CommonServiceLocator.SolrNet/IContainer.cs
