+++ NHibernate.SolrNet/NHibernate.SolrNet.nuspec
@@ -12,7 +12,7 @@
     <description>NHibernate integration for SolrNet</description>
     <language>en-US</language>
     <dependencies>
-      <dependency id="SolrNet" version="[0.7.1]" />
+      <dependency id="SolrNet.Core" version="[0.7.1]" />
       <dependency id="NHibernate" version="3.2.0.4000" />
     </dependencies>
   </metadata>
