+++ Ninject.Integration.SolrNet/Ninject.Integration.SolrNet.nuspec
@@ -12,7 +12,7 @@
     <description>Ninject module for SolrNet</description>
     <language>en-US</language>
     <dependencies>
-      <dependency id="SolrNet" version="[0.7.1]" />
+      <dependency id="SolrNet.Core" version="[0.7.1]" />
       <dependency id="Ninject" version="3.2.2.0" />
     </dependencies>
   </metadata>
