+++ SolrNet.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
-VisualStudioVersion = 15.0.26730.3
+VisualStudioVersion = 15.0.26730.8
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{26EBA068-9BBB-4FEC-B425-787D4AE99848}"
 	ProjectSection(SolutionItems) = preProject
@@ -48,6 +48,10 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = ".nuget", ".nuget", "{648832
 	ProjectSection(SolutionItems) = preProject
 		.nuget\NuGet.Config = .nuget\NuGet.Config
 	EndProjectSection
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CommonServiceLocator.SolrNet", "CommonServiceLocator.SolrNet\CommonServiceLocator.SolrNet.csproj", "{17B34150-5102-435B-A487-84DD5803A228}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CommonServiceLocator.SolrNet.Tests", "CommonServiceLocator.SolrNet.Tests\CommonServiceLocator.SolrNet.Tests.csproj", "{0988D6A2-04FD-4328-AFFA-677DB377432D}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -123,6 +127,14 @@ Global
 		{4BC966D4-2EC0-47FF-B784-B649BFAABD81}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{4BC966D4-2EC0-47FF-B784-B649BFAABD81}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{4BC966D4-2EC0-47FF-B784-B649BFAABD81}.Release|Any CPU.Build.0 = Release|Any CPU
+		{17B34150-5102-435B-A487-84DD5803A228}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{17B34150-5102-435B-A487-84DD5803A228}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{17B34150-5102-435B-A487-84DD5803A228}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{17B34150-5102-435B-A487-84DD5803A228}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0988D6A2-04FD-4328-AFFA-677DB377432D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0988D6A2-04FD-4328-AFFA-677DB377432D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0988D6A2-04FD-4328-AFFA-677DB377432D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0988D6A2-04FD-4328-AFFA-677DB377432D}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
