+++ SolrNet/SolrNet.csproj
@@ -72,9 +72,6 @@
     </AssemblyOriginatorKeyFile>
   </PropertyGroup>
   <ItemGroup>
-    <Reference Include="Microsoft.Practices.ServiceLocation, Version=1.3.0.0, Culture=neutral, PublicKeyToken=31bf3856ad364e35, processorArchitecture=MSIL">
-      <HintPath>..\packages\CommonServiceLocator.1.3\lib\portable-net4+sl5+netcore45+wpa81+wp8\Microsoft.Practices.ServiceLocation.dll</HintPath>
-    </Reference>
     <Reference Include="System" />
     <Reference Include="System.configuration" />
     <Reference Include="System.Data" />
@@ -374,9 +371,7 @@
     <Compile Include="Impl\SolrServer.cs" />
     <Compile Include="SortOrder.cs" />
     <Compile Include="StartOrCursor.cs" />
-    <Compile Include="Startup.cs" />
     <Compile Include="StatsResult.cs" />
-    <Compile Include="Utils\Container.cs" />
     <Compile Include="Exceptions\InvalidURLException.cs" />
     <Compile Include="ISolrConnection.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
@@ -385,7 +380,6 @@
     <Compile Include="Exceptions\SolrNetException.cs" />
     <Compile Include="Utils\HttpEncoder.cs" />
     <Compile Include="Utils\HttpUtility.cs" />
-    <Compile Include="Utils\IContainer.cs" />
     <Compile Include="Utils\KV.cs" />
     <Compile Include="Utils\Memoizer.cs" />
     <Compile Include="Utils\TypeHelper.cs" />
