+++ SolrNet/SolrNet.nuspec
@@ -1,9 +1,9 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <package xmlns="http://schemas.microsoft.com/packaging/2010/07/nuspec.xsd">
     <metadata>
-        <id>SolrNet</id>
+        <id>SolrNet.Core</id>
         <version>0.7.1</version>
-        <title>SolrNet</title>
+        <title>SolrNet.Core</title>
         <authors>Mauricio Scheffer and contributors</authors>
         <owners>Mauricio Scheffer</owners>
         <licenseUrl>https://raw.githubusercontent.com/SolrNet/SolrNet/master/license.txt</licenseUrl>
@@ -11,9 +11,6 @@
         <requireLicenseAcceptance>false</requireLicenseAcceptance>
         <description>Apache Solr client</description>
         <language>en-US</language>
-        <dependencies>
-            <dependency id="CommonServiceLocator" version="1.3" />
-        </dependencies>
     </metadata>
     <files>
         <file src="bin\release\SolrNet.dll" target="lib\net46\SolrNet.dll" />
