+++ Unity.SolrNetIntegration/Unity.SolrNetIntegration.nuspec
@@ -12,7 +12,7 @@
     <description>Unity integration for SolrNet</description>
     <language>en-US</language>
     <dependencies>
-      <dependency id="SolrNet" version="[0.7.1]" />
+      <dependency id="SolrNet.Core" version="[0.7.1]" />
       <dependency id="Unity" version="2.1.505" />
     </dependencies>
   </metadata>
