--- packages/packages.config
@@ -1,4 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Fuchu" version="0.4.0" targetFramework="net40" />
-</packages>
