--- packages/repositories.config
@@ -1,11 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<repositories>
-  <repository path="..\AutofacContrib.SolrNet.Tests\packages.config" />
-  <repository path="..\AutofacContrib.SolrNet\packages.config" />
-  <repository path="..\Castle.Facilities.SolrNetIntegration.Tests\packages.config" />
-  <repository path="..\Castle.Facilities.SolrNetIntegration\packages.config" />
-  <repository path="..\Ninject.Integration.SolrNet.Tests\packages.config" />
-  <repository path="..\Ninject.Integration.SolrNet\packages.config" />
-  <repository path="..\SolrNet.Tests\packages.config" />
-  <repository path="packages.config" />
-</repositories>
